fix some bugs and refactor rip.py
