+++ dvd-ripper/tests/test_parse_makemkv_info.py
@@ -2,7 +2,7 @@
 
 import unittest
 
-from rip import parse_makemkv_info
+from titles import parse_makemkv_info
 from tests.test_data import AVATAR_DISC_INFO, MOVIE_DISC_INFO
